+++ tests/test_prg_engine_control_flow_support.h
@@ -287,6 +287,7 @@ void test_case_predicates_use_heap_backed_expression_checkpoints();
 void test_loop_predicates_and_bounds_use_heap_backed_expression_checkpoints();
 void test_scan_predicate_preserves_rest_scope_and_exhaustion_state();
 void test_cancel_releases_frame_owned_native_objects();
+void test_cancel_releases_frame_owned_private_native_objects();
 void test_elseif_predicate_resumption_review_gaps();
 void test_expression_level_function_call_can_chain_nested_user_routines();
 void test_set_procedure_registers_external_function_for_expression_calls();